pyproject.toml
[tool.poetry]
name = "eagletrt-telemetria-validator"
version = "0.1.0"
description = ""
authors = ["nicolatoscan <nicolatoscan99@outlook.it>"]

[tool.poetry.dependencies]
python = "^3.8"
python-can = "^3.3.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
